+++ AllaganItemSearch/AllaganItemSearchPlugin.cs
@@ -1,29 +1,36 @@
+using System;
 using System.Globalization;
 using System.Reflection;
 
+using AllaganItemSearch.Boot;
 using AllaganItemSearch.Filters;
 using AllaganItemSearch.ItemRenderers;
 using AllaganItemSearch.Services;
 using AllaganItemSearch.Settings;
 using AllaganItemSearch.Settings.Layout;
 using AllaganItemSearch.Windows;
+
 using AllaganLib.Data.Service;
 using AllaganLib.GameSheets.Extensions;
+using AllaganLib.GameSheets.Modules;
+using AllaganLib.GameSheets.Service;
 using AllaganLib.Interface.Services;
 using AllaganLib.Interface.Widgets;
 using AllaganLib.Interface.Wizard;
 using AllaganLib.Shared.Time;
 
 using Autofac;
+
 using DalaMock.Host.Hosting;
-using DalaMock.Host.Mediator;
 using DalaMock.Shared.Classes;
 using DalaMock.Shared.Interfaces;
+
 using Dalamud.Game.Text;
 using Dalamud.Interface.ImGuiFileDialog;
 using Dalamud.Interface.Windowing;
 using Dalamud.Plugin;
 using Dalamud.Plugin.Services;
+
 using Lumina.Excel;
 using Lumina.Excel.Sheets;
 
@@ -35,6 +42,8 @@ namespace AllaganItemSearch;
 
 public class AllaganItemSearchPlugin : HostedPlugin
 {
+    private readonly BootConfigurationService bootService;
+
     public AllaganItemSearchPlugin(
         IDalamudPluginInterface pluginInterface,
         IPluginLog pluginLog,
@@ -70,15 +79,16 @@ public class AllaganItemSearchPlugin : HostedPlugin
             gameGui,
             keyState)
     {
+        this.bootService = new BootConfigurationService(pluginInterface, framework, pluginLog);
         this.CreateHost();
         this.Start();
     }
 
     public override HostedPluginOptions ConfigureOptions()
     {
-        return new HostedPluginOptions()
+        return new HostedPluginOptions
         {
-            UseMediatorService = true
+            UseMediatorService = true,
         };
     }
 
@@ -86,16 +96,18 @@ public class AllaganItemSearchPlugin : HostedPlugin
     {
         var dataAccess = Assembly.GetExecutingAssembly();
 
-        containerBuilder.Register(
-            s =>
-            {
-                // Assume we only ever have one number format but we could just make this keyed later
-                var gilNumberFormat =
-                    (NumberFormatInfo)CultureInfo.CurrentCulture.NumberFormat.Clone();
-                gilNumberFormat.CurrencySymbol = SeIconChar.Gil.ToIconString();
-                gilNumberFormat.CurrencyDecimalDigits = 0;
-                return gilNumberFormat;
-            });
+        containerBuilder.Register(s =>
+        {
+            // Assume we only ever have one number format but we could just make this keyed later
+            var gilNumberFormat =
+                (NumberFormatInfo)CultureInfo.CurrentCulture.NumberFormat.Clone();
+            gilNumberFormat.CurrencySymbol = SeIconChar.Gil.ToIconString();
+            gilNumberFormat.CurrencyDecimalDigits = 0;
+            return gilNumberFormat;
+        });
+
+        containerBuilder.RegisterInstance(this.bootService).ExternallyOwned().AsSelf();
+        containerBuilder.RegisterInstance(this.bootService.Configuration).ExternallyOwned().AsSelf();
 
         containerBuilder.RegisterAssemblyTypes(dataAccess)
                         .Where(t => t.Name.EndsWith("Setting"))
@@ -132,7 +144,13 @@ public class AllaganItemSearchPlugin : HostedPlugin
         containerBuilder.RegisterType<ClipboardService>().As<IClipboardService>().SingleInstance();
         containerBuilder.RegisterType<TryOnService>().As<ITryOnService>().SingleInstance();
         containerBuilder.RegisterType<ATService>().SingleInstance();
-        containerBuilder.RegisterGameSheetManager();
+        containerBuilder.RegisterModule(new GameSheetManagerModule()
+        {
+            StartupOptions = new SheetManagerStartupOptions()
+            {
+                PersistInDataShare = this.bootService.Configuration.PersistLuminaCache,
+            },
+        });
 
         containerBuilder.RegisterType<FileDialogManager>().SingleInstance();
         containerBuilder.RegisterType<DalamudFileDialogManager>().As<IFileDialogManager>().SingleInstance();
@@ -142,21 +160,22 @@ public class AllaganItemSearchPlugin : HostedPlugin
         containerBuilder.RegisterType<ConfigWindow>().As<Window>().AsSelf().SingleInstance();
         containerBuilder.RegisterType<WizardWindow>().As<Window>().AsSelf().SingleInstance();
 
-        //Hotkey System
-        containerBuilder.RegisterType<HotkeyService<Configuration>>().AsSelf().AsImplementedInterfaces().SingleInstance();
+        // Hotkey System
+        containerBuilder.RegisterType<HotkeyService<Configuration>>().AsSelf().AsImplementedInterfaces()
+                        .SingleInstance();
 
         containerBuilder.Register(c => c.Resolve<IDataManager>().GameData).SingleInstance();
 
         // Custom Widgets
-        containerBuilder.Register(
-            c => new WizardWidgetSettings() { PluginName = "Allagan Item Search", LogoPath = "logo_small" });
+        containerBuilder.Register(c => new WizardWidgetSettings
+        { PluginName = "Allagan Item Search", LogoPath = "logo_small" });
         containerBuilder.RegisterType<WizardWidget<Configuration>>().AsSelf().AsImplementedInterfaces()
                         .SingleInstance();
         containerBuilder.RegisterType<ConfigurationWizardService<Configuration>>().AsSelf().AsImplementedInterfaces()
                         .SingleInstance();
         containerBuilder.RegisterType<Font>().As<IFont>().SingleInstance();
 
-        //Filters
+        // Filters
         containerBuilder.RegisterType<FilterService>().SingleInstance();
         containerBuilder.RegisterType<StringFilter>();
         containerBuilder.RegisterType<YesNoChoiceFilter>();
@@ -172,35 +191,30 @@ public class AllaganItemSearchPlugin : HostedPlugin
                         .As<IItemInfoRenderer>()
                         .SingleInstance();
 
-
         // Sheets
-        containerBuilder.Register<ExcelSheet<Item>>(
-            s =>
-            {
-                var dataManger = s.Resolve<IDataManager>();
-                return dataManger.GetExcelSheet<Item>()!;
-            }).SingleInstance();
+        containerBuilder.Register<ExcelSheet<Item>>(s =>
+        {
+            var dataManger = s.Resolve<IDataManager>();
+            return dataManger.GetExcelSheet<Item>()!;
+        }).SingleInstance();
 
-        containerBuilder.Register<ExcelSheet<ClassJob>>(
-            s =>
-            {
-                var dataManger = s.Resolve<IDataManager>();
-                return dataManger.GetExcelSheet<ClassJob>()!;
-            }).SingleInstance();
+        containerBuilder.Register<ExcelSheet<ClassJob>>(s =>
+        {
+            var dataManger = s.Resolve<IDataManager>();
+            return dataManger.GetExcelSheet<ClassJob>()!;
+        }).SingleInstance();
 
-        containerBuilder.Register<ExcelSheet<World>>(
-            s =>
-            {
-                var dataManger = s.Resolve<IDataManager>();
-                return dataManger.GetExcelSheet<World>()!;
-            }).SingleInstance();
+        containerBuilder.Register<ExcelSheet<World>>(s =>
+        {
+            var dataManger = s.Resolve<IDataManager>();
+            return dataManger.GetExcelSheet<World>()!;
+        }).SingleInstance();
 
-        containerBuilder.Register(
-            s =>
-            {
-                var configurationLoaderService = s.Resolve<ConfigurationLoaderService>();
-                return configurationLoaderService.GetConfiguration();
-            }).SingleInstance();
+        containerBuilder.Register(s =>
+        {
+            var configurationLoaderService = s.Resolve<ConfigurationLoaderService>();
+            return configurationLoaderService.GetConfiguration();
+        }).SingleInstance();
     }
 
     public override void ConfigureServices(IServiceCollection serviceCollection)
@@ -210,10 +224,24 @@ public class AllaganItemSearchPlugin : HostedPlugin
         serviceCollection.AddHostedService(p => p.GetRequiredService<WindowService>());
         serviceCollection.AddHostedService(p => p.GetRequiredService<CommandService>());
         serviceCollection.AddHostedService(p => p.GetRequiredService<InstallerWindowService>());
-        serviceCollection.AddHostedService(p => p.GetRequiredService<MediatorService>());
         serviceCollection.AddHostedService(p => p.GetRequiredService<LaunchButtonService>());
         serviceCollection.AddHostedService(p => p.GetRequiredService<ConfigurationLoaderService>());
         serviceCollection.AddHostedService(p => p.GetRequiredService<AutoSaveService>());
         serviceCollection.AddHostedService(p => p.GetRequiredService<ATService>());
+    }
+
+    protected virtual void Dispose(bool disposing)
+    {
+        if (disposing)
+        {
+            this.bootService.Dispose();
+        }
+    }
+
+    public new void Dispose()
+    {
+        base.Dispose();
+        this.Dispose(true);
+        GC.SuppressFinalize(this);
     }
 }